backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain-openai
langchain-core
pydantic
spacy